ha-box/requirements.txt
# Communication avec Home Assistant
requests>=2.31.0

# Matériel - Bus de communication
smbus2>=0.4.3
spidev>=3.6
RPi.GPIO>=0.7.1

# Matériel - Périphériques
adafruit-circuitpython-bme280>=2.5.11
adafruit-circuitpython-pn532>=1.4.0

# Graphisme
Pillow>=10.0.0

# Utilitaires
python-dateutil>=2.8.2
